Remove unecessary casts

+++ src/main/java/com/hotpads/datarouter/client/imp/hbase/cluster/DRHRegionInfo.java
@@ -131,7 +131,7 @@ implements Comparable<DRHRegionInfo<?>>{
 		String totalKvsString = DrNumberFormatter.addCommas(totalKvs);
 		long compactingKvs = load.getCurrentCompactedKVs();
 		if(totalKvs==compactingKvs){ return totalKvsString; }
-		int percentCompacted = (int)((double)100 * (double)compactingKvs / (double)totalKvs);
+		int percentCompacted = (int)((double)100 * (double)compactingKvs / totalKvs);
 		return totalKvsString + " ["+percentCompacted+"%]";
 	}
